add logic into moveToken method

+++ src/main/kotlin/com/snakesandladders/domain/MoveTokenData.kt
@@ -1,3 +1,3 @@
 package com.snakesandladders.domain
 
-data class MoveTokenData(var game: BoardGame, var moves: Long, var playerId: Long)
+data class MoveTokenData(var game: BoardGame, var moves: Int, var playerId: Long)
